refactor: move `ProjectSeed` model from `core` app into `project` app
- Moved `ProjectSeed` model to models file in `project` app.
- Created initial migration for the `ProjectSeed` model ownership from `core` to `project` app using `SeparateDatabaseAndState` to update migration state and rename `core_projectseed` to `project_projectseed`. Also update content type before changing the state from `ContentType("core", "projectseed")` to `ContentType("project", "projectseed")`
- Updated imports in various files to reflect the new location of the `ProjectSeed` model.
- Modified tests to import the ProjectSeed model from its new location.
- Refactor imports and references in previous migration files to reflect the change from `core.project` to `project.project` for `ProjectSeed` model and methods
