+++ docker-app/qfieldcloud/core/admin.py
@@ -67,7 +67,6 @@ from qfieldcloud.core.models import (
     Person,
     ProcessProjectfileJob,
     ProjectCollaborator,
-    ProjectSeed,
     Secret,
     Team,
     TeamMember,
@@ -80,7 +79,11 @@ from qfieldcloud.core.utils2 import delta_utils, jobs, pg_service_file
 from qfieldcloud.core.utils2.storage import format_storage_usage
 from qfieldcloud.filestorage.backend import QfcS3Boto3Storage
 from qfieldcloud.filestorage.models import File
-from qfieldcloud.project.models import SHARED_DATASETS_PROJECT_NAME, Project
+from qfieldcloud.project.models import (
+    SHARED_DATASETS_PROJECT_NAME,
+    Project,
+    ProjectSeed,
+)
 from qfieldcloud.subscription.models import get_subscription_model
